hmr fix + removal of forgoten debug statements
