+++ src/Bridge/Symfony/Bundle/SwooleBundle.php
@@ -5,7 +5,9 @@ declare(strict_types=1);
 namespace K911\Swoole\Bridge\Symfony\Bundle;
 
 use K911\Swoole\Bridge\Symfony\Bundle\DependencyInjection\CompilerPass\EntityManagerDecoratorPass;
-use K911\Swoole\Bridge\Symfony\Bundle\DependencyInjection\CompilerPass\GoAopAspectKernelOverridePass;
+use K911\Swoole\Server\Runtime\HMR\HmrComposerLoader;
+use K911\Swoole\Server\Runtime\HMR\LoadedFiles;
+use Symfony\Component\Debug\DebugClassLoader;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\HttpKernel\Bundle\Bundle;
 
@@ -19,6 +21,47 @@ final class SwooleBundle extends Bundle
     public function build(ContainerBuilder $container): void
     {
         $container->addCompilerPass(new EntityManagerDecoratorPass());
-        $container->addCompilerPass(new GoAopAspectKernelOverridePass());
+    }
+
+    /**
+     *
+     */
+    public function boot()
+    {
+        // it is a quick way to check if loader was enabled
+        $wasDebugEnabled = class_exists(DebugClassLoader::class, false);
+        if ($wasDebugEnabled) {
+            // disable temporary to apply AOP loader first
+            DebugClassLoader::disable();
+        }
+
+        $this->initHmrComposerLoaders();
+
+        if ($wasDebugEnabled) {
+            DebugClassLoader::enable();
+        }
+    }
+
+    /**
+     * @return void
+     */
+    private function initHmrComposerLoaders(): void
+    {
+        $loaders = spl_autoload_functions();
+
+        $files = $this->container->get(LoadedFiles::class);
+
+        foreach ($loaders as &$loader) {
+            $loaderToUnregister = $loader;
+            if (is_array($loader) && $loader[0]) {
+                $loader[0] = new HmrComposerLoader($files, $loader[0]);
+            }
+            spl_autoload_unregister($loaderToUnregister);
+        }
+        unset($loader);
+
+        foreach ($loaders as $loader) {
+            spl_autoload_register($loader);
+        }
     }
 }
